chore: split base entity file
