+++ apps/ai-dial-admin/src/components/ActivityAudit/View/utils.ts
@@ -1,7 +1,7 @@
 import { ModelViewI18nKey } from '@/src/constants/i18n';
 import { NO_LIMITS_KEY } from '@/src/constants/role';
 import { ActivityAuditDiff, ActivityAuditSection } from '@/src/models/activity-audit';
-import { DialRoleLimits } from '@/src/models/dial/base-entity';
+import { DialRoleLimits } from '@/src/models/dial/role-limits';
 import { DialModelEndpoint, DialModelPricing, PricingType } from '@/src/models/dial/model';
 import { EntitiesGridData } from '@/src/models/entities-grid-data';
 import { ActivityAuditEntity, ActivityAuditResourceType, DiffStatus } from '@/src/types/activity-audit';
