+++ apps/ai-dial-admin/src/components/AddEntitiesTab/AddEntitiesView.utils.ts
@@ -4,7 +4,8 @@ import { ENTITY_BASE_COLUMNS } from '@/src/constants/grid-columns/grid-columns';
 import { LIMIT_COLUMNS } from '@/src/components/EntityView/Roles/utils';
 import { NO_LIMITS_KEY } from '@/src/constants/role';
 import { DialApplication, DialApplicationScheme } from '@/src/models/dial/application';
-import { DialBaseEntity, DialRoleShare } from '@/src/models/dial/base-entity';
+import { DialRoleShare } from '@/src/models/dial/role-limits';
+import { DialBaseEntity } from '@/src/models/dial/base-entity';
 import { DialInterceptor } from '@/src/models/dial/interceptor';
 import { DialKey } from '@/src/models/dial/key';
 import { DialModel } from '@/src/models/dial/model';
