+++ apps/ai-dial-admin/src/components/EntityView/Features/Features.tsx
@@ -4,7 +4,8 @@ import Switch from '@/src/components/Common/Switch/Switch';
 import EndpointControl from '@/src/components/EntityMainProperties/BaseProperties/Endpoint/Endpoint';
 import { FeaturesI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
-import { DialBaseEntity, DialFeatures } from '@/src/models/dial/base-entity';
+import { DialBaseEntity } from '@/src/models/dial/base-entity';
+import { DialFeatures } from '@/src/models/dial/features';
 import { ApplicationRoute } from '@/src/types/routes';
 import { placeholdersMap } from './constants';
 import { getSwitchControls, getTextControls } from './utils';
