+++ apps/ai-dial-admin/src/components/EntityView/Roles/RolesDefaults.tsx
@@ -3,7 +3,8 @@ import { FC, useCallback } from 'react';
 import { NumberInputField } from '@/src/components/Common/InputField/InputField';
 import { EntityFieldsI18nKey, RolesI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
-import { DialBaseEntity, DialRoleLimits, DialRoleShare } from '@/src/models/dial/base-entity';
+import { DialBaseEntity } from '@/src/models/dial/base-entity';
+import { DialRoleLimits, DialRoleShare } from '@/src/models/dial/role-limits';
 
 interface Props {
   entity: DialBaseEntity;
