+++ apps/ai-dial-admin/src/components/EntityView/Roles/models.ts
@@ -1,4 +1,4 @@
-import { DialRoleLimits, DialRoleShare } from '@/src/models/dial/base-entity';
+import { DialRoleLimits, DialRoleShare } from '@/src/models/dial/role-limits';
 
 export interface RolesGridData extends DialRoleLimits, DialRoleShare {
   name?: string;
