+++ apps/ai-dial-admin/src/models/dial/default.ts
@@ -1,0 +1,6 @@
+export interface DefaultTemp {
+  key: string;
+  value: DefaultsValue;
+}
+
+export type DefaultsValue = string | number | boolean;
