+++ apps/ai-dial-admin/src/models/dial/features.ts
@@ -1,0 +1,20 @@
+export interface DialFeatures {
+  truncatePromptEndpoint: string;
+  rateEndpoint: string;
+  configurationEndpoint: string;
+  tokenizeEndpoint: string;
+  toolsSupported: boolean;
+  systemPromptSupported: boolean;
+  urlAttachmentsSupported: boolean;
+  folderAttachmentsSupported: boolean;
+  seedSupported: boolean;
+  allowResume: boolean;
+  addonsSupported: boolean;
+  temperatureSupported: boolean;
+  accessibleByPerRequestKey: boolean;
+  parallelToolCallsSupported: boolean;
+  contentPartsSupported: boolean;
+  cacheSupported: boolean;
+  autoCachingSupported: boolean;
+  consentRequired: boolean;
+}
