+++ apps/ai-dial-admin/src/models/dial/interceptor.ts
@@ -1,5 +1,6 @@
-import { DefaultsValue, DefaultTemp, DialBaseNamedEntity } from './base-entity';
+import { DialBaseNamedEntity } from './base-entity';
 import { SOURCE_TYPE } from '@/src/components/SourceField/types';
+import { DefaultsValue, DefaultTemp } from '@/src/models/dial/default';
 
 export interface DialInterceptor extends DialBaseNamedEntity {
   endpoint?: string;
