+++ apps/ai-dial-admin/src/models/dial/role-limits.ts
@@ -1,0 +1,14 @@
+export type DialRoleLimitsMap = Record<string, DialRoleLimits>;
+export type DialRoleShareMap = Record<string, DialRoleShare>;
+
+export interface DialRoleLimits {
+  day?: string | null;
+  minute?: string | null;
+  week?: string | null;
+  month?: string | null;
+}
+
+export interface DialRoleShare {
+  invitationTtl?: string | null;
+  maxAcceptedUsers?: string | null;
+}
