+++ apps/ai-dial-admin/src/models/dial/role.ts
@@ -1,4 +1,5 @@
-import { DialBaseNamedEntity, DialRoleLimits, DialRoleShare } from './base-entity';
+import { DialBaseNamedEntity } from './base-entity';
+import { DialRoleLimits, DialRoleShare } from '@/src/models/dial/role-limits';
 
 export interface DialRole extends DialBaseNamedEntity {
   limits?: Record<string, DialRoleLimits>;
